// back-end/src/constants/enum.ts
enum TokenType {
  AccessToken,
  RefreshToken,
  ForgotPasswordToken,
  EmailVerifyToken
}

enum USERS_MESSAGES {
  REFRESH_TOKEN_IS_REQUIRED = 'REFRESH TOKEN IS REQUIRED',
  USED_REFRESH_TOKEN_OR_NOT_EXISTED = 'REFRESH TOKEN IS NOT EXISTED',
  REFRESH_TOKEN_IS_INVALID = 'REFRESH TOKEN IS INVALID',
  EMAIL_TOKEN_IS_REQUIRED = 'EMAIL TOKEN IS REQUIRED',
  USER_IS_NOT_EXISTED = 'USER IS NOT EXISTED',
  USER_HAS_ALREADY_VERIFIED = 'USER HAS ALREADY VERIFIED',
  VERIFY_EMAIL_SUCCESS = 'VERIFY EMAIL SUCCESS',
  RESEND_EMAIL_VERIFY_SUCCESS = 'RESEND EMAIL SUCCESS',
  SEND_TOKEN_RESET_FORGOT_PASSWORD = 'SEND RESET PASSWORD TOKEN SUCCESSFULLY',
  ACCESS_TOKEN_IS_REQUIRED = 'ACCESS TOKEN IS REQUIRED'
}

enum HTTP_STATUS {
  UNAUTHORIZED = 401,
  NOT_FOUND = 404,
  SUCCESS = 200
}

export { TokenType, USERS_MESSAGES, HTTP_STATUS }